+++ src/stuart-ml/util/isInstanceOf.lua
@@ -1,3 +1,0 @@
-local isInstanceOf = function(obj, class)
-  return obj.isInstanceOf ~= nil and obj:isInstanceOf(class) end
-return isInstanceOf