+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 ### Added
+- `clustering:` Port `KMeansModel`
+- `util:` Port `Loader`, which can load a `KMeansModel` from a Parquet file or directory of files, whether local and WebHDFS
 - `util:` Port `NumericParser`
 - `util:` Port `StringTokenizer`